rolled back changes

+++ controller/AdministrationController.php
@@ -15,33 +15,45 @@ require_once  __DIR__ . '/../config/Picnic.php';
 class AdministrationController {
 
 	/**
-	 * Displays the user administration page.
+	 * Displays the main user administration page.
 	 */
 	public function Users(): void {
-		$this->verifyAuthorization();
+		if ($this->auth())  {
+			$h = new Humphree(Picnic::getInstance());
 
-		$view = new AdminUsersView();
-		$view->Render('adminUsers');
+			$pagerData = Pager::ParsePagerDataFromQuery();
+			$pagerData->totalItems = sizeof($h->getUsers(1, 1000000)); // temporary, until we get a countUsers() method.
+
+			$view = new View();
+			$view->SetData('navData', new NavData(NavData::Account));
+			$view->SetData('users', $h->getUsers($pagerData->pageNumber, $pagerData->itemsPerPage));
+			$view->SetData('pagerData', $pagerData);
+			$view->Render('administration');
+		} else {
+			$this->handleUnauthorised();
+		}
 	}
 
 	/**
-	 * Displays the category administration page.
+	 * Displays a page for changing the given user's password.
+	 *
+	 * @param int $userID		The ID of the user whose password will be changed.
 	 */
-	public function ViewCategories(): void {
-		$this->verifyAuthorization();
+	public function ChangePassword(int $userID) {
+		if ($this->auth()) {
+			if ($_SERVER['REQUEST_METHOD'] === 'GET') {
+				$view = new View();
+				$view->SetData('navData', new NavData(NavData::Account));
+				$view->SetData('userID', $userID);
+				$view->Render('adminChangePassword');
+				return;
+			} else if (isset ($_POST ['update'])) {
+				$this->doChangePassword($userID);
+				return;
+			}
+		}
 
-		$view = new AdminCategoriesView();
-		$view->Render('adminCategories');
-	}
-
-	/**
-	 * Displays the system administration page.
-	 */
-	public function System(): void {
-		$this->verifyAuthorization();
-
-		$view = new AdminSystemView();
-		$view->Render('adminSystem');
+		header('Location: ' . BASE . '/Home');
 	}
 
 	/**
@@ -51,15 +63,33 @@ class AdministrationController {
 	 * 			The ID of the user to be edited.
 	 */
 	public function EditUser(int $userID): void {
-		$this->verifyAuthorization();
-		$this->verifyValidUserId($userID);
+		if ($this->auth()) {
+			if ($_SERVER['REQUEST_METHOD'] === 'GET') {
+				$user = new User (Picnic::getInstance());
+				$user->userID = $userID;
 
-		if (isset($_POST['update'])) {
-			$this->updateUser($userID);
-		} else {
-			$view = new AdminEditUserView($userID);
-			$view->Render('adminEditUser');
+				if ($user->get()) {
+
+					$userData = [
+						'userID' => $user->userID,
+						'user' => $user->user,
+						'email' => $user->email,
+						'status' => $user->status
+					];
+
+					$view = new View();
+					$view->SetData('navData', new NavData(NavData::Account));
+					$view->SetData('user', $userData);
+					$view->Render('adminEditUser');
+					return;
+				}
+			} else if (isset ($_POST ['update'])) {
+				$this->updateUser($userID);
+				return;
+			}
 		}
+
+		header('Location: ' . BASE . '/Home');
 	}
 
 	/**
@@ -69,11 +99,13 @@ class AdministrationController {
 	 * 			The ID of the user to be deleted.
 	 */
 	public function DeleteUser(int $userID): void {
-		$this->verifyAuthorization();
-
-		$h = new Humphree(Picnic::getInstance());
-		$h->deleteUser($userID);
-		header('Location: ' . BASE . '/Administration/Users');
+		if ($this->auth()) {
+			$h = new Humphree(Picnic::getInstance());
+			$h->deleteUser($userID);
+			header('Location: ' . BASE . '/Administration/Users');
+		} else {
+			header('Location: ' . BASE . '/Home');
+		}
 	}
 
 	/**
@@ -83,11 +115,15 @@ class AdministrationController {
 	 * 			The ID of the user to be blocked.
 	 */
 	public function BlockUser(int $userID): void {
-		$this->verifyAuthorization();
-
-		$h = new Humphree(Picnic::getInstance());
-		$h->blockUser($userID);
-		header('Location: ' . BASE . '/Administration/Users');
+		if ($this->auth()) {
+			$h = new Humphree(Picnic::getInstance());
+			if (!$h->blockUser($userID)) {
+				$_SESSION['error'] = 'Failed to block the user.';
+			}
+			header('Location: ' . BASE . '/Administration/Users');
+		} else {
+			header('Location: ' . BASE . '/Home');
+		}
 	}
 
 	/**
@@ -97,84 +133,52 @@ class AdministrationController {
 	 * 			The ID of the user to be unblocked.
 	 */
 	public function UnblockUser(int $userID): void {
-		$this->verifyAuthorization();
+		if ($this->auth()) {
+			$h = new Humphree(Picnic::getInstance());
+			if (!$h->unblockUser($userID)) {
+				$_SESSION['error'] = 'Failed to unblock the user.';
+			}
 
-		$h = new Humphree(Picnic::getInstance());
-		$h->unblockUser($userID);
-		header('Location: ' . BASE . '/Administration/Users');
-	}
-
-	/**
-	 * Displays a page for changing the given user's password.
-	 *
-	 * @param int $userID		The ID of the user whose password will be changed.
-	 */
-	public function ChangePassword(int $userID): void {
-		$this->verifyAuthorization();
-		$this->verifyValidUserId($userID);
-
-		if ($_SERVER['REQUEST_METHOD'] === 'GET') {
-			$view = new View();
-			$view->SetData('navData', new NavData(NavData::Account));
-			$view->SetData('userID', $userID);
-			$view->Render('adminChangePassword');
-		} else if (isset($_POST['update'])) {
-			$this->doChangePassword($userID);
-		} else {
 			header('Location: ' . BASE . '/Administration/Users');
+		} else {
+			header('Location: ' . BASE . '/Home');
 		}
 	}
 
-	/**
-	 * Adds a new top-level item category.
-	 */
+	public function ViewCategories(): void {
+		$view = new AdminCategoriesView();
+		$view->Render('adminCategories');
+		return;
+	}
+
 	public function AddMajorCategory(): void {
-		$this->verifyAuthorization();
-
-		if (isset($_POST['category'])) {
-			$h = new Humphree(Picnic::getInstance());
-			$categoryName = $_POST['category'];
-			$h->addCategory(Category::ROOT_CATEGORY, $categoryName);
-		}
+		$h = new Humphree(Picnic::getInstance());
+		$categoryName = $_POST['category'];
+		$h->addCategory(Category::ROOT_CATEGORY, $categoryName);
 		header('Location: ' . BASE . '/Administration/ViewCategories');
 	}
 
-	/**
-	 * Adds a new second-level item category.
-	 *
-	 * @param int $parentCategoryID		The ID of the top-level category under
-	 * 									which the new category will be created.
-	 */
 	public function AddMinorCategory(int $parentCategoryID): void {
-		$this->verifyAuthorization();
-
-		if (isset($_POST['category'])) {
-			$h = new Humphree(Picnic::getInstance());
-			$categoryName = $_POST['category'];
-			$h->addCategory($parentCategoryID, $categoryName);
-		}
+		$h = new Humphree(Picnic::getInstance());
+		$categoryName = $_POST['category'];
+		$h->addCategory($parentCategoryID, $categoryName);
 		header('Location: ' . BASE . '/Administration/ViewCategories');
 	}
 
-	/**
-	 * Deletes the category with the given ID.
-	 *
-	 * @param int $categoryID	The ID of the category to be deleted.
-	 */
 	public function DeleteCategory(int $categoryID): void {
-		$this->verifyAuthorization();
-
 		$h = new Humphree(Picnic::getInstance());
 		$h->deleteCategory($categoryID);
 		header('Location: ' . BASE . '/Administration/ViewCategories');
 	}
 
-	/**
-	 * Deletes the database and recreates it with default content.
-	 */
-	public function RebuildDatabase() {
-		$this->verifyAuthorization();
+	public function System() {
+		$view = new AdminSystemView();
+		$view->Render('adminSystem');
+		return;
+	}
 
+	public function RebuildDatabase()
+	{
 		echo 'Creating database...<br />';
 		flush();
 
@@ -220,56 +224,87 @@ class AdministrationController {
 		$view->Render('createDB');
 	}
 
+	private function auth(){
+		return isset($_SESSION[MODULE])
+			&& isset($_SESSION['userID'])
+			&& isset($_SESSION['status'])
+			&& $_SESSION['userID'] > 0
+			&& $_SESSION['status'] === 'admin';
+	}
+
 	/**
 	 * Updates the user attributes if they are valid, then displays the main admin page.
-	 *
-	 * @param int $userID	The ID of the user to be updated.
+	 * @param int $userID
 	 */
 	private function updateUser(int $userID): void {
-		if ($this->validatePostedUserInfo()) {
-			$h = new Humphree(Picnic::getInstance());
-			$user = $h->getUser($userID);
-			$user['user'] = $_POST ['user'];
-			$user['email'] = $_POST ['email'];
-			$user['status'] = $_POST ['status'];
+		if ($this->auth()) {
+			$user = new User (Picnic::getInstance());
+			$v = new Validation ();
 
-			try {
-				$h->updateUser($user);
-			} catch (ModelException $e) {
-				// temporary -- to catch the case where no change was made to the data, in which
-				// case update throws due to no rows being modified.
+			// Process on submission of user.
+			if (isset ($_POST ['user'])) {
+
+				// Validate the user.
+				try {
+					$v->userName($_POST ['user']);
+				} catch (ValidationException $e) {
+					$_SESSION ['error'] = $e->getError();
+				}
+
+				if (isset ($_SESSION ['error'])) {
+					unset ($_POST ['user']);
+					header('Location: Administration');
+				} else {
+					$user->user = $_POST ['user'];
+					unset ($_POST ['user']);
+
+					// Validate the email.
+					try {
+						$v->email($_POST ['email']);
+					} catch (ValidationException $e) {
+						$_SESSION ['error'] = $e->getError();
+					}
+
+					if (isset ($_SESSION ['error'])) {
+						unset ($_POST ['email']);
+						header('Location: Administration');
+					} else {
+						$user->email = $_POST ['email'];
+						unset ($_POST ['email']);
+						$user->status = $_POST ['status'];
+						unset ($_POST ['status']);
+						$user->userID = $userID;
+						try {
+							$user->update();
+						} catch (ModelException $e) {
+							// temporary -- to catch the case where no change was made to the data, in which
+							// case update throws due to now rows being modified.
+						}
+
+
+						header('Location: ' . BASE . '/Administration/Users');
+					}
+				}
 			}
-
-			header('Location: ' . BASE . '/Administration/Users');
 		} else {
-			header('Location: ' . BASE . '/Administration/EditUser/' . $userID);
+			header('Location: ' . BASE . '/Home');
 		}
 	}
 
 	/**
-	 * Updates the user password if it is valid. Displays the main user admin page
-	 * if successful, otherwise returns to the change password page.
-	 *
-	 * @param int $userID	The ID of the user to be updated.
+	 * Updates the user password if it is valid, then displays the main admin page.
+	 * @param int $userID
 	 */
 	private function doChangePassword(int $userID): void {
-		$user = new User (Picnic::getInstance());
-		$v = new Validation ();
+		if ($this->auth()) {
+			$user = new User (Picnic::getInstance());
+			$v = new Validation ();
 
-		if (isset ($_POST ['password1']) && isset ($_POST ['password2'])) {
-			try {
-				$v->password($_POST ['password1']);
-			} catch (ValidationException $e) {
-				$_SESSION ['error'] = $e->getError();
-			}
+			if (isset ($_POST ['password1']) && isset ($_POST ['password2'])) {
 
-			if (isset ($_SESSION ['error'])) {
-				unset ($_POST ['password1']);
-				unset ($_POST ['password2']);
-				header('Location: ' . BASE . '/Administration/ChangePassword/' . $userID);
-			} else {
+				// Validate the password.
 				try {
-					$v->comparePasswords($_POST ['password1'], $_POST ['password2']);
+					$v->password($_POST ['password1']);
 				} catch (ValidationException $e) {
 					$_SESSION ['error'] = $e->getError();
 				}
@@ -277,81 +312,37 @@ class AdministrationController {
 				if (isset ($_SESSION ['error'])) {
 					unset ($_POST ['password1']);
 					unset ($_POST ['password2']);
-					header('Location: Administration');
+					header('Location: ' . BASE . '/Administration/ChangePassword/' . $userID);
 				} else {
-					$user->password = $_POST ['password1'];
-					unset ($_POST ['password1']);
-					unset ($_POST ['password2']);
-					$user->userID = $userID;
-					$user->updatePassword();
 
-					header('Location: ' . BASE . '/Administration/Users');
+					// Compare passwords.
+					try {
+						$v->comparePasswords($_POST ['password1'], $_POST ['password2']);
+					} catch (ValidationException $e) {
+						$_SESSION ['error'] = $e->getError();
+					}
+
+					if (isset ($_SESSION ['error'])) {
+						unset ($_POST ['password1']);
+						unset ($_POST ['password2']);
+						header('Location: Administration');
+					} else {
+						$user->password = $_POST ['password1'];
+						unset ($_POST ['password1']);
+						unset ($_POST ['password2']);
+						$user->userID = $userID;
+						$user->updatePassword();
+
+						header('Location: ' . BASE . '/Administration/Users');
+					}
 				}
 			}
-		}
-	}
-
-	/**
-	 * Validates that all posted user data is valid.
-	 *
-	 * @return bool		True if the data is valid, false otherwise.
-	 */
-	private function validatePostedUserInfo(): bool
-	{
-		try {
-			$v = new Validation ();
-
-			if (!isset($_POST['user'])) {
-				throw new ValidationException('Name cannot be empty.');
-			}
-
-			$v->userName($_POST['user']);
-
-			if (!isset($_POST['email'])) {
-				throw new ValidationException('Email cannot be empty.');
-			}
-
-			$v->email($_POST ['email']);
-
-			if (!isset($_POST['email'])) {
-				throw new ValidationException('Email cannot be empty.');
-			}
-
-			$v->email($_POST ['email']);
-
-			return true;
-		} catch (ValidationException $e) {
-			$_SESSION ['error'] = $e->getError();
-			return false;
-		}
-	}
-
-	/**
-	 * Confirms that the current user is logged in as an administrator. If they are not,
-	 * they are redirected to the home page.
-	 */
-	private function verifyAuthorization(): void{
-		if (!isset($_SESSION[MODULE])
-			|| !isset($_SESSION['userID'])
-			|| !isset($_SESSION['status'])
-			|| $_SESSION['userID'] <= 0
-			|| $_SESSION['status'] !== 'admin') {
+		} else {
 			header('Location: ' . BASE . '/Home');
-			die();
 		}
 	}
 
-	/**
-	 * Confirms the a user exits with the given ID. If not, the user is redirected
-	 * to the admin users page.
-	 *
-	 * @param int $userID	The user ID to be tested.
-	 */
-	private function verifyValidUserId(int $userID): void {
-		$h = new Humphree(Picnic::getInstance());
-		if (!$h->isValidUserID($userID)) {
-			header('Location: ' . BASE . '/Administration/Users');
-			die();
-		}
+	private function handleUnauthorised(): void {
+		header('Location: ' . BASE . '/Home');
 	}
 }
